int -> size_t

+++ Jolt/Core/TickCounter.cpp
@@ -56,7 +56,7 @@ static uint64 sProcessorTicksPerSecond = []() {
 		#endif
 
 			// Check if line starts with correct string
-			const int num_chars = strlen(cpu_str);
+			const size_t num_chars = strlen(cpu_str);
 			if (strncmp(line.c_str(), cpu_str, num_chars) == 0)
 			{
 				// Find ':'
